# Conflicts:
#	grails-doc/src/en/guide/upgrading/upgrading80x.adoc

+++ grails-data-hibernate7/core/src/main/groovy/org/grails/orm/hibernate/HibernateGormStaticApi.groovy
@@ -309,9 +309,9 @@ class HibernateGormStaticApi<D> extends GormStaticApi<D> {
     }
 
     @Override
-    Integer count() {
+    Long count() {
         String entity = persistentEntity.name
-        doSingleInternal("select count(*) from $entity" as String, [:], [], [:], false) as Integer
+        doSingleInternal("select count(*) from $entity" as String, [:], [], [:], false) as Long
     }
 
     @Override
